Refactor geodesics into free function
